% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcs_confusion.R
\name{npcs_confusion}
\alias{npcs_confusion}
\title{Fit a multi-class Neyman-Pearson classifier with error controls via cost-sensitive learning.}
\usage{
npcs_confusion(
  x,
  y,
  algorithm = c("CX", "ER"),
  classifier = c("logistic", "knn", "randomforest", "tree", "neuralnet", "svm", "lda",
    "qda", "nb", "nnb", "logistic-l1"),
  w,
  alpha,
  split.ratio = 0.5,
  split.mode = c("by-class", "merged"),
  tol = 1e-06,
  refit = TRUE,
  protect = TRUE,
  opt.alg = c("Hooke-Jeeves", "Nelder-Mead"),
  limit = 200,
  fitted.model = NULL,
  inf.threshold = 1,
  ...
)
}
\description{
Fit a multi-class Neyman-Pearson classifier with error controls via cost-sensitive learning.
}
