% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.npcs.R
\name{predict.npcs}
\alias{predict.npcs}
\title{Predict new labels from new data based on the fitted NPMC classifier.}
\usage{
\method{predict}{npcs}(object, newx, ...)
}
\arguments{
\item{object}{the model object for prediction}

\item{newx}{input feature data}

\item{...}{arguments to pass down}
}
\description{
Predict new labels from new data based on the fitted NPMC classifier, which belongs to S3 class "npcs".
}
