% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.npcs.R
\name{print.cv.npcs}
\alias{print.cv.npcs}
\title{Print the cv.npcs object.}
\usage{
\method{print}{cv.npcs}(x, ...)
}
\arguments{
\item{x}{fitted cv.npcs object using \code{cv.npcs}.}

\item{...}{additional arguments.}
}
\description{
Print the cv.npcs object.
}
